refactor: reorganize source file handling modules
