+++ src/inspector/index.ts
@@ -11,3 +11,4 @@ export type {
 } from "./inspector.ts";
 export type { InspectorResult, InspectorResults } from "./inspector-result.ts";
 export { runInspectors } from "./runner.ts";
+export type { FileType, ParsedSourceFile, ParsedSourceFileMetadata } from "./source-file-types.ts";
