+++ src/inspector/runner.ts
@@ -3,9 +3,9 @@
  */
 
 import ts from "typescript";
-import type { ParsedSourceFile } from "../source-file/index.ts";
 import type { FileInspectionResult, Inspector } from "./inspector.ts";
 import type { InspectorResults } from "./inspector-result.ts";
+import type { ParsedSourceFile } from "./source-file-types.ts";
 
 /**
  * Inspects all the given source files using the provided inspectors,
