+++ src/main/orchestrator/source-file/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Source file parsing implementation for orchestrator.
+ */
+
+export { inferParseSourceFilesOptions, type ParseSourceFilesOptions } from "./options.ts";
+export { parseSourceFiles } from "./parser.ts";
